pyproject.toml
[project]
name = "WWW2026_OPL"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.6"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "obp>=0.4.1",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
]
